% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{string_reverse}
\alias{string_reverse}
\title{Reverses a string or number}
\usage{
string_reverse(x, convert = TRUE)
}
\arguments{
\item{x}{A string or a number}

\item{convert}{Flag to set whether the input number will be converted to strings or not}
}
\value{
the reverse of input
}
\description{
Reverses a string or number
}
\examples{
string_reverse(1234)
string_reverse("1234")
}
